Refactor: GeneralCode에서 CommonResponseCode로 변경
