+++ be/src/main/java/codesquad/secondhand/controller/ItemController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import static codesquad.secondhand.exception.code.GeneralCode.RESPONSE_SUCCESS;
+import static codesquad.secondhand.exception.code.CommonResponseCode.RESPONSE_SUCCESS;
 @Slf4j
 @RestController
 @RequiredArgsConstructor
